+++ tokenspeed-kernel/python/tokenspeed_kernel/ops/moe/__init__.py
@@ -29,7 +29,6 @@ import tokenspeed_kernel.ops.moe.cuda  # noqa: F401
 import tokenspeed_kernel.ops.moe.deepep  # noqa: F401
 import tokenspeed_kernel.ops.moe.flashinfer  # noqa: F401
 import tokenspeed_kernel.ops.moe.gluon  # noqa: F401
-import tokenspeed_kernel.ops.moe.gluon_decode_routing_gfx950  # noqa: F401
 import tokenspeed_kernel.ops.moe.triton  # noqa: F401
 import tokenspeed_kernel.ops.moe.triton_kernels  # noqa: F401
 import tokenspeed_kernel.ops.moe.trtllm  # noqa: F401
