+++ tokenspeed-kernel/test/ops/moe/test_gluon_decode_routing_gfx950.py
@@ -2,10 +2,10 @@ from __future__ import annotations
 
 import pytest
 import tokenspeed_kernel  # noqa: F401  (registers moe kernels)
-import tokenspeed_kernel.ops.moe.gluon_decode_routing_gfx950 as gluon_mod
+import tokenspeed_kernel.ops.moe.gluon as gluon_mod
 import torch
 from tokenspeed_kernel.ops.moe import moe_route
-from tokenspeed_kernel.ops.moe.gluon_decode_routing_gfx950 import (
+from tokenspeed_kernel.ops.moe.gluon import (
     SMALLM_MAX_M,
     gluon_fused_route,
     gluon_route_supported,
